refactor(react-markdown): single-source the code-fence contract on a subpath
