+++ packages/react-markdown/package.json
@@ -22,6 +22,17 @@
     "./styles/dot.css": {
       "style": "./styles/dot.css",
       "default": "./styles/dot.css"
+    },
+    "./code-fence": {
+      "types": "./dist/code-fence.d.ts",
+      "default": "./dist/code-fence.js"
+    }
+  },
+  "typesVersions": {
+    "*": {
+      "code-fence": [
+        "./dist/code-fence.d.ts"
+      ]
     }
   },
   "main": "./dist/index.js",
@@ -41,6 +52,7 @@
   "dependencies": {
     "@radix-ui/react-primitive": "^2.1.10",
     "@radix-ui/react-use-callback-ref": "^1.1.4",
+    "@types/hast": "^3.0.5",
     "classnames": "^2.5.1",
     "react-markdown": "^10.1.0"
   },
@@ -57,7 +69,6 @@
   "devDependencies": {
     "@assistant-ui/react": "workspace:*",
     "@assistant-ui/x-buildutils": "workspace:*",
-    "@types/hast": "^3.0.5",
     "@types/react": "^19.2.18",
     "react": "^19.2.8",
     "react-dom": "^19.2.8",
