+++ packages/react-markdown/src/overrides/types.ts
@@ -1,25 +1,6 @@
-import type { Element } from "hast";
-import type { ComponentPropsWithoutRef, ComponentType } from "react";
-
-export type PreComponent = ComponentType<
-  ComponentPropsWithoutRef<"pre"> & { node?: Element | undefined }
->;
-export type CodeComponent = ComponentType<
-  ComponentPropsWithoutRef<"code"> & { node?: Element | undefined }
->;
-
-export type CodeHeaderProps = {
-  node?: Element | undefined;
-  language: string | undefined;
-  code: string;
-};
-
-export type SyntaxHighlighterProps = {
-  node?: Element | undefined;
-  components: {
-    Pre: PreComponent;
-    Code: CodeComponent;
-  };
-  language: string;
-  code: string;
-};
+export type {
+  CodeComponent,
+  CodeHeaderProps,
+  PreComponent,
+  SyntaxHighlighterProps,
+} from "../code-fence";
